refactor: exception폴더 global내부로 이동
