+++ src/main/java/com/project/coalba/domain/substituteReq/service/BossSubstituteReqService.java
@@ -7,8 +7,8 @@ import com.project.coalba.domain.substituteReq.dto.response.YearMonth;
 import com.project.coalba.domain.substituteReq.entity.SubstituteReq;
 import com.project.coalba.domain.substituteReq.repository.SubstituteRepository;
 import com.project.coalba.domain.substituteReq.repository.dto.BothSubstituteReqDto;
-import com.project.coalba.exception.BusinessException;
-import com.project.coalba.exception.ErrorCode;
+import com.project.coalba.global.exception.BusinessException;
+import com.project.coalba.global.exception.ErrorCode;
 import com.project.coalba.global.utils.ProfileUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
