+++ src/main/java/com/project/coalba/domain/substituteReq/service/StaffSubstituteReqService.java
@@ -12,8 +12,8 @@ import com.project.coalba.domain.substituteReq.entity.enums.SubstituteReqStatus;
 import com.project.coalba.domain.substituteReq.repository.dto.BothSubstituteReqDto;
 import com.project.coalba.domain.substituteReq.repository.SubstituteRepository;
 import com.project.coalba.domain.substituteReq.repository.dto.SubstituteReqDto;
-import com.project.coalba.exception.BusinessException;
-import com.project.coalba.exception.ErrorCode;
+import com.project.coalba.global.exception.BusinessException;
+import com.project.coalba.global.exception.ErrorCode;
 import com.project.coalba.global.utils.ProfileUtil;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
