+++ src/main/java/com/project/coalba/global/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package com.project.coalba.exception;
+package com.project.coalba.global.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
