+++ src/main/java/com/project/coalba/global/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.project.coalba.exception;
+package com.project.coalba.global.exception;
 
 import lombok.Builder;
 import lombok.Getter;
