+++ src/main/java/com/project/coalba/global/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.project.coalba.exception;
+package com.project.coalba.global.exception;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
